Switch to JPA using MySQL
This commit showcases what is required to migrate the fst demo from
Mongo to JPA using MySQL. These are the few required changes:

1. The relevant spring-data-jpa project should be used instead of the
mongo integration.
2. The driver of the database should be added.
3. A few properties defining the URL of the database and the username
password are required to use the proper database.

Note that no JPA or Hibernate configuration is required. As we scan for
repositories to turn them into Mongo-based implementations, switching
to spring-data-jpa scans them as well as any entity defined in the
package (or sub-package) of the application by default.

Note also that the driver of the database is not required as Boot
detects it from the jdbc url.